+++ python/pytauri/src/pytauri_plugins/notification/__init__.py
@@ -1,13 +1,15 @@
 """[tauri_plugin_notification::self](https://docs.rs/tauri-plugin-notification/latest/tauri_plugin_notification/index.html)"""
 
-from pytauri_plugin_notification.ffi import (
+from pytauri_plugins.notification.ffi import (
     ImplNotificationExt,
     NotificationBuilder,
+    NotificationBuilderArgs,
     NotificationExt,
 )
 
 __all__ = [
     "ImplNotificationExt",
     "NotificationBuilder",
+    "NotificationBuilderArgs",
     "NotificationExt",
 ]