+++ python/pytauri/src/pytauri_plugins/notification/ffi.py
@@ -6,31 +6,52 @@
 """
 
 from types import ModuleType
-from typing import TYPE_CHECKING, Optional, final
+from typing import TYPE_CHECKING, final
 
-from pytauri import EXT_MOD, ImplManager
-from typing_extensions import TypeAlias
+from pytauri import ImplManager
+from typing_extensions import TypeAlias, TypedDict, Unpack
+
+from pytauri_plugins import (
+    PLUGIN_NOTIFICATION,
+    _pytauri_plugins_mod,  # pyright: ignore[reportPrivateUsage]
+)
 
 __all__ = [
     "ImplNotificationExt",
     "NotificationBuilder",
+    "NotificationBuilderArgs",
     "NotificationExt",
 ]
 
-
-def _load_notification_mod(ext_mod: ModuleType) -> ModuleType:
-    try:
-        notification_mod = ext_mod.notification
-    except AttributeError as e:
-        raise RuntimeError(
-            "Submodule `notification` is not found in the extension module"
-        ) from e
-
-    assert isinstance(notification_mod, ModuleType)
-    return notification_mod
+if PLUGIN_NOTIFICATION:
+    _notification_mod: ModuleType = _pytauri_plugins_mod.notification
+else:
+    raise ImportError(
+        "Enable the `plugin-notification` feature for `pytauri` crate to use this plugin."
+    )
 
 
-_notification_mod = _load_notification_mod(EXT_MOD)
+class NotificationBuilderArgs(TypedDict, total=False):
+    """[tauri_plugin_notification::NotificationBuilder](https://docs.rs/tauri-plugin-notification/latest/tauri_plugin_notification/struct.NotificationBuilder.html)"""
+
+    id: int
+    channel_id: str
+    title: str
+    body: str
+    large_body: str
+    summary: str
+    action_type_id: str
+    group: str
+    group_summary: bool
+    sound: str
+    inbox_line: str
+    icon: str
+    large_icon: str
+    icon_color: str
+    ongoing: bool
+    auto_cancel: bool
+    silent: bool
+
 
 if TYPE_CHECKING:
 
@@ -38,28 +59,7 @@ if TYPE_CHECKING:
     class NotificationBuilder:
         """[tauri_plugin_notification::NotificationBuilder](https://docs.rs/tauri-plugin-notification/latest/tauri_plugin_notification/struct.NotificationBuilder.html)"""
 
-        def show(
-            self,
-            /,
-            *,
-            id: Optional[int] = None,  # noqa: A002
-            channel_id: Optional[str] = None,
-            title: Optional[str] = None,
-            body: Optional[str] = None,
-            large_body: Optional[str] = None,
-            summary: Optional[str] = None,
-            action_type_id: Optional[str] = None,
-            group: Optional[str] = None,
-            group_summary: bool = False,
-            sound: Optional[str] = None,
-            inbox_line: Optional[str] = None,
-            icon: Optional[str] = None,
-            large_icon: Optional[str] = None,
-            icon_color: Optional[str] = None,
-            ongoing: bool = False,
-            auto_cancel: bool = False,
-            silent: bool = False,
-        ) -> None:
+        def show(self, /, **kwargs: Unpack[NotificationBuilderArgs]) -> None:
             """Consume this builder and show the notification.
 
             # FIXME, XXX, TODO:
@@ -86,4 +86,4 @@ else:
     NotificationExt = _notification_mod.NotificationExt
 
 ImplNotificationExt: TypeAlias = ImplManager
-"""The implementors of `NotificationExt`."""
+"""The implementers of `NotificationExt`."""